requirements.txt
# Core dependencies
neo4j>=5.0.0
click>=8.1.3
pydantic>=2.0.0
typer>=0.9.0
rich>=13.0.0
streamlit>=1.25.0
faiss-cpu>=1.7.0
sentence-transformers>=2.0.0
openai>=1.0.0
python-dotenv>=1.0.0
langchain>=0.1.0
langchain-community>=0.0.10
langchain-openai>=0.0.10
langchain-google-genai>=2.1.2
jinja2>=3.1.2

# Development dependencies
pytest>=7.4.0
black>=23.7.0
isort>=5.12.0
flake8>=6.1.0 